docs: major refactor to include updates to alice_bob_model and corresponding tutorials
