+++ docs/tutorial/etl.md
@@ -33,7 +33,8 @@ In this case, instead of using a custom extract class to define the data source,
 As always, we begin by adding a new file to the `generators` module and define a function that accepts the model as an input then begins with `with model:` then `with Generator(name=...):`
 
 ```python3
-{!../docs_src/tutorials/alice-bob-model/alice_bob_model/generators/f_to_c.py [ln:18-20] !}
+{!../examples/alice_bob_lab/{{cookiecutter.repo_name}}/alice_bob_model/generators/f_to_c.py [ln:14-17] !}
+            ...
 ```
 
 ## Query
@@ -48,8 +49,8 @@ For queries that include only one table (which is the case in this example), the
 
 Similarly to the extracts shown in the previous generators, by calling `.results()` on the `Query`, a tuple of the outputs are returned in the order that they are listed in the `select` statement.
 
-```python3 hl_lines="11-13"
-{!../docs_src/tutorials/alice-bob-model/alice_bob_model/generators/f_to_c.py [ln:1-2,4,18-] !}
+```python3 hl_lines="8-10"
+{!../examples/alice_bob_lab/{{cookiecutter.repo_name}}/alice_bob_model/generators/f_to_c.py [ln:2-3,15-20] !}
 ```
 
 ## Imports and Environments
@@ -82,8 +83,8 @@ Essentially, whenever a line like...
 
 Finally, it is worth noting that dbgen `Env` objects can be added together. So, if there is a default `Env` that is used for most transforms, and we just need to add one extra import to that, rather than define a new `Env` from scratch that includes every package, we can simply create an `Env` with just the one new import and add it to the default `Env`. An example of doing just that is shown below.
 
-```python3 hl_lines="8"
-{!../docs_src/tutorials/alice-bob-model/alice_bob_model/generators/f_to_c.py [ln:1-] !}
+```python3
+{!../examples/alice_bob_lab/{{cookiecutter.repo_name}}/alice_bob_model/generators/f_to_c.py [ln:1-] !}
 ```
 
 
@@ -91,8 +92,8 @@ Finally, it is worth noting that dbgen `Env` objects can be added together. So, 
 
 Once the environment is defined, the transform step is very similar to the previously-shown transform steps. We define a function that carries out the desired data analysis and add the `@transform` decorator to specify the output names and the python environment.
 
-```python3 hl_lines="7-15"
-{!../docs_src/tutorials/alice-bob-model/alice_bob_model/generators/f_to_c.py [ln:1-] !}
+```python3
+{!../examples/alice_bob_lab/{{cookiecutter.repo_name}}/alice_bob_model/generators/f_to_c.py [ln:1-] !}
 ```
 
 ## Load
@@ -109,6 +110,6 @@ And, as always, when `.load()` is called, we must supply either:
     - calling `Tablename.load()`, which always returns the id of the specified row (as shown in the previous generator).
 
 
-```python3 hl_lines="25-28"
-{!../docs_src/tutorials/alice-bob-model/alice_bob_model/generators/f_to_c.py [ln:1-] !}
+```python3
+{!../examples/alice_bob_lab/{{cookiecutter.repo_name}}/alice_bob_model/generators/f_to_c.py [ln:1-] !}
 ```
