+++ docs/tutorial/io_gen_complex.md
@@ -19,7 +19,7 @@
 Oftentimes, scientific data is stored in custom file formats, and we need to be able to ingest this into the database. In this example, we assume there is a folder of temperature measurements on the local file system. The file names are of the format `FirstName_LastName_MeasurementNumber.txt`, and each file contains text similar to the following:
 
 ```
-{!../docs_src/tutorials/alice-bob-model/data/measurements/Alice_Smith_0.txt!}
+{!../examples/alice_bob_lab/{{cookiecutter.repo_name}}/data/measurements/Alice_Smith_0.txt!}
 ```
 
 # Writing the Custom Extract
@@ -28,8 +28,8 @@ Oftentimes, scientific data is stored in custom file formats, and we need to be 
 
 First, we subclass the Extract class (imported from dbgen) and define the necessary inputs.
 
-```python3 hl_lines="9-12"
-{!../docs_src/tutorials/alice-bob-model/alice_bob_model/extracts/measurement_extract.py [ln:1-12] !}
+```python3
+{!../examples/alice_bob_lab/{{cookiecutter.repo_name}}/alice_bob_model/extracts/measurement_extract.py [ln:1-26] !}
 ```
 
 The `data_dir` is a string that specifies the name of the folder in which our text files are stored.
@@ -42,8 +42,8 @@ Finally, we want to be able to read in the list of filenames in the `data_dir` i
 
 Next, we need to overwrite the `setup` method. In this case, all we'll do is use `os.listdir` to get a list of the filenames in the folder that contains our text files.
 
-```python3 hl_lines="14-15"
-{!../docs_src/tutorials/alice-bob-model/alice_bob_model/extracts/measurement_extract.py [ln:1-15] !}
+```python3
+{!../examples/alice_bob_lab/{{cookiecutter.repo_name}}/alice_bob_model/extracts/measurement_extract.py [ln:1-30] !}
 ```
 
 ## Defining the extract method
@@ -52,8 +52,8 @@ Finally, we need to define the `extract` method. The extract method must always 
 
 We loop over the filenames stored in our private attribute, and simply read the file and return a dictionary containing the file name and file contents.
 
-```python3 hl_lines="17-22"
-{!../docs_src/tutorials/alice-bob-model/alice_bob_model/extracts/measurement_extract.py !}
+```python3
+{!../examples/alice_bob_lab/{{cookiecutter.repo_name}}/alice_bob_model/extracts/measurement_extract.py !}
 ```
 
 Parsing these strings will happen in the transform step.
@@ -65,39 +65,39 @@ The first step to defining a transform is to define the output names and the pyt
 
 We want our function to return four items: the researchers first name and last name, the order in which the measurement was taken, and the actual temperature measurement.
 
-```python3 hl_lines="9-13"
-{!../docs_src/tutorials/alice-bob-model/alice_bob_model/generators/parse_measurements.py [ln:1-20] !}
+```python3
+{!../examples/alice_bob_lab/{{cookiecutter.repo_name}}/alice_bob_model/generators/parse_measurements.py [ln:1-10] !}
 ```
 
 Next, we need to write a custom function that parses the filename and file contents to extract the information that we are interested in.
 
 
-```python3 hl_lines="14-20"
-{!../docs_src/tutorials/alice-bob-model/alice_bob_model/generators/parse_measurements.py [ln:1-20] !}
+```python3
+{!../examples/alice_bob_lab/{{cookiecutter.repo_name}}/alice_bob_model/generators/parse_measurements.py [ln:1-32] !}
 ```
 
 # Inserting the values into the database
 
 Next, in order to insert these values into the database, we need to define a dbgen Generator. The standard pattern is to define a function that accepts the model and adds the new generator to the model. By using `with model`, all Generators defined in that with block will automatically be added to the model. Similarly, by using `with Generator(...)`, all extracts, transforms, and loads instantiated in that with block will automatically be added to the new Generator. Lines similar to the ones highlighted below are used almost every time a new Generator is defined.
 
-```python3 hl_lines="1-3"
-{!../docs_src/tutorials/alice-bob-model/alice_bob_model/generators/parse_measurements.py [ln:23-] !}
+```python3  hl_lines="2-3"
+{!../examples/alice_bob_lab/{{cookiecutter.repo_name}}/alice_bob_model/generators/parse_measurements.py [ln:31-] !}
 ```
 
 Next, we need to instantiate the custom extract we defined above. By calling `.results()` on the instance of our custom extract class, a tuple of the outputs is returned.
 
 ```python3 hl_lines="4"
-{!../docs_src/tutorials/alice-bob-model/alice_bob_model/generators/parse_measurements.py [ln:23-] !}
+{!../examples/alice_bob_lab/{{cookiecutter.repo_name}}/alice_bob_model/generators/parse_measurements.py [ln:31-] !}
 ```
 
 After that, we want to pass the results from this extract to our custom transform (the parser defined above). Similarly, by calling `.results()` on the transform, a tuple of the outputs is returned.
 
 ```python3 hl_lines="5"
-{!../docs_src/tutorials/alice-bob-model/alice_bob_model/generators/parse_measurements.py [ln:23-] !}
+{!../examples/alice_bob_lab/{{cookiecutter.repo_name}}/alice_bob_model/generators/parse_measurements.py [ln:31-] !}
 ```
 
 Finally, we call `.load(...)` on the table that we would like to insert data into, and we pass the values output by the transform as keyword arguments to the `.load` method. An important point is that any call to `.load` returns the ID of the row specified in the `.load(...)` statement. We do not always need to use this information, but we do need it to populate foreign keys. Simply put, foreign keys are always populated by calling the `.load` method on the table that you would like to create a foreign key to, as shown in the last line below.
 
 ```python3 hl_lines="6-11"
-{!../docs_src/tutorials/alice-bob-model/alice_bob_model/generators/parse_measurements.py [ln:23-] !}
+{!../examples/alice_bob_lab/{{cookiecutter.repo_name}}/alice_bob_model/generators/parse_measurements.py [ln:31-] !}
 ```
