+++ examples/alice_bob_lab/{{cookiecutter.repo_name}}/alice_bob_model/generators/f_to_c.py
@@ -1,31 +1,13 @@
-#   Copyright 2021 Modelyst LLC
-#
-#   Licensed under the Apache License, Version 2.0 (the "License");
-#   you may not use this file except in compliance with the License.
-#   You may obtain a copy of the License at
-#
-#       http://www.apache.org/licenses/LICENSE-2.0
-#
-#   Unless required by applicable law or agreed to in writing, software
-#   distributed under the License is distributed on an "AS IS" BASIS,
-#   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#   See the License for the specific language governing permissions and
-#   limitations under the License.
-
 from alice_bob_model.constants import DEFAULT_ENV
 from alice_bob_model.schema import TemperatureMeasurement
-from scipy.constants import convert_temperature
 from sqlmodel import select
 
-from dbgen import Env, Generator, Import, Model, Query, transform
-
-outputs = ["temp_c"]
-env = DEFAULT_ENV + Env([Import("scipy.constants", "convert_temperature")])
+from dbgen import Generator, Model, Query, transform
 
 
-@transform(outputs=outputs, env=env)
+@transform(outputs=["temp_c"], env=DEFAULT_ENV)
 def f_to_c(temp_f: float) -> float:
-    temp_c = convert_temperature(temp_f, "F", "C")
+    temp_c = (temp_f - 32.0) * 5.0 / 9.0
 
     return temp_c
