+++ examples/alice_bob_lab/{{cookiecutter.repo_name}}/alice_bob_model/generators/parse_measurements.py
@@ -1,17 +1,4 @@
-#   Copyright 2021 Modelyst LLC
-#
-#   Licensed under the Apache License, Version 2.0 (the "License");
-#   you may not use this file except in compliance with the License.
-#   You may obtain a copy of the License at
-#
-#       http://www.apache.org/licenses/LICENSE-2.0
-#
-#   Unless required by applicable law or agreed to in writing, software
-#   distributed under the License is distributed on an "AS IS" BASIS,
-#   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#   See the License for the specific language governing permissions and
-#   limitations under the License.
-
+import re
 from os.path import join
 from typing import Tuple
 
@@ -19,18 +6,26 @@ from alice_bob_model.constants import DATA_DIR, DEFAULT_ENV
 from alice_bob_model.extracts.measurement_extract import MeasurementExtract
 from alice_bob_model.schema import Person, TemperatureMeasurement
 
-from dbgen import Generator, Model, transform
-
-outputs = ["first_name", "last_name", "ordering", "temperature"]
-env = DEFAULT_ENV
+from dbgen import Env, Generator, Import, Model, transform
 
 
-@transform(outputs=outputs, env=env)
-def parse_measurements(file_name: str, contents: str) -> Tuple[str, str, float]:
-    file_name = file_name.strip(".txt")
-    first_name, last_name, ordering_str = file_name.split("_")
+@transform(
+    outputs=["first_name", "last_name", "ordering", "temperature"], env=DEFAULT_ENV + Env(Import("re"))
+)
+def parse_measurements(file_name: str, contents: str) -> Tuple[str, str, int, float]:
+    regex = r"([A-Za-z]+)_([A-Za-z]+)_(\d+).txt"
+    match = re.match(regex, file_name)
+    if not match:
+        raise ValueError("No match found")
+    first_name, last_name, ordering_str = match.groups()
     ordering = int(ordering_str)
-    temperature = float(contents.split(": ")[-1])
+
+    regex = r".*:\s*(\d+)"
+    match = re.match(regex, contents)
+    if not match:
+        raise ValueError("No match found")
+    (temperature_str,) = match.groups()
+    temperature = float(temperature_str)
 
     return first_name, last_name, ordering, temperature
