+++ examples/alice_bob_lab/{{cookiecutter.repo_name}}/alice_bob_model/schema.py
@@ -1,21 +1,6 @@
-#   Copyright 2021 Modelyst LLC
-#
-#   Licensed under the Apache License, Version 2.0 (the "License");
-#   you may not use this file except in compliance with the License.
-#   You may obtain a copy of the License at
-#
-#       http://www.apache.org/licenses/LICENSE-2.0
-#
-#   Unless required by applicable law or agreed to in writing, software
-#   distributed under the License is distributed on an "AS IS" BASIS,
-#   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#   See the License for the specific language governing permissions and
-#   limitations under the License.
-
 from typing import Optional
-from uuid import UUID
 
-from dbgen import Entity
+from dbgen import Entity, IDType
 
 
 class Person(Entity, table=True):
@@ -32,4 +17,4 @@ class TemperatureMeasurement(Entity, table=True):
     temperature_F: Optional[float]
     temperature_C: Optional[float]
     ordering: Optional[int]
-    person_id: UUID = Person.foreign_key()
+    person_id: IDType = Person.foreign_key()
